chore: moved the mocking server logic to be shared by all the tests
